fix: 修复控制中心调整连接方向不生效
参数类型错误，改为quint16类型

Log: 控制中心调整连接方向接口参数修改为quint16类型
Influence: 显示模块跨端协同方向设置涉及功能及流程
Bug: https://pms.uniontech.com/bug-view-179251.html

+++ src/plugin-display/operation/machine.cpp
@@ -93,7 +93,7 @@ void Machine::setUUID(const QString &uuid)
     m_UUID = uuid;
 }
 
-void Machine::setDirection(const int dir)
+void Machine::setDirection(const quint16 dir)
 {
     if (m_direction != dir) {
         m_direction = dir;
